+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## vNext
 ### Added
 - ThreadNumber is emitted in the result
+- Extension to log to log to the console 
   
 ## v2.2.1
 ### Fixed